Further reduce test memory usage

+++ server/src/test/java/umm3601/admin/TestQRCodes.java
@@ -74,13 +74,12 @@ public class TestQRCodes {
 
     @Test
     public void TestCreateBufferedImages() throws IOException,WriterException{
-        String bedURLs[] = new String[2];
+        String bedURLs[] = new String[1];
         bedURLs[0]= "http://localhost:2538/bed/bed1?qr=true";
-        bedURLs[1] = "http://localhost:2538/bed/bed2?qr=true";
         System.out.println(QRCodes.createBufferedImages(bedURLs).get(0));
 
         List<BufferedImage> qrCodeImages = new ArrayList<BufferedImage>();
-        for(int i = 0; i < 2; i++) {
+        for(int i = 0; i < 1; i++) {
             qrCodeImages.add(createQRFromBedURL(bedURLs[i]));
         }
 
@@ -90,16 +89,14 @@ public class TestQRCodes {
 
     @Test
     public void TestWriteZipPathForQRCodes() throws IOException,WriterException{
-        String bedNames[] = new String[2];
+        String bedNames[] = new String[1];
         bedNames[0] = "bed1";
-        bedNames[1] = "bed2";
 
-        String bedURLs[] = new String[2];
+        String bedURLs[] = new String[1];
         bedURLs[0] = "http://localhost:2538/bed/bed1?qr=true";
-        bedURLs[1] = "http://localhost:2538/bed/bed2?qr=true";
         List<BufferedImage> qrCodeImages = QRCodes.createBufferedImages(bedURLs);
 
-        assertEquals(qrCodeImages.size(), 2);
+        assertEquals(qrCodeImages.size(), 1);
 
         QRCodes.writeBufferedImagesToFile(qrCodeImages,bedNames,this.path);
 
@@ -125,20 +122,22 @@ public class TestQRCodes {
             //Delete temp folder holding QRCodes
             File f = new File(path);
             String files[] = f.list();
-            for (int i = 0; i < files.length; i++) {
-                Files.delete(Paths.get(path + "/" + files[i]));
-            }
+            if(files != null) {
+                for (int i = 0; i < files.length; i++) {
+                    Files.delete(Paths.get(path + "/" + files[i]));
+                }
 
-            Path tempFolderPath = Paths.get(this.path);
-            if (Files.exists(tempFolderPath))
-                Files.delete(tempFolderPath);
+                Path tempFolderPath = Paths.get(this.path);
+                if (Files.exists(tempFolderPath))
+                    Files.delete(tempFolderPath);
 
-            //Delete QRCode zip file
-            f = new File("./");
-            files = f.list();
-            for (int i = 0; i < files.length; i++) {
-                if (files[i].endsWith(".zip") && files[i].startsWith("QR Code Export"))
-                    Files.delete(Paths.get(files[i]));
+                //Delete QRCode zip file
+                f = new File("./");
+                files = f.list();
+                for (int i = 0; i < files.length; i++) {
+                    if (files[i].endsWith(".zip") && files[i].startsWith("QR Code Export"))
+                        Files.delete(Paths.get(files[i]));
+                }
             }
         }
         catch(Exception e)
